totally rewrite scraper using capybara
It seems that there's some sort of session magic occuring which capybara
handles much better so use that instead of nokogiri.

This now works by creating a memberships table into which it puts every
historical term/person combination it finds. It then uses that to scrape
things into data, updating the memberships table with any new
historical memberships it finds.

It also populates the terms table as it goes with names and dates
although this does mean the source URL is a bit unhelpful.

If it's already seen a person or a membership then it doesn't scrape
them again unless either the RESCRAPE_ALL env is set or the
RESCRAPE_TERM env variable is set to the term of the membership.

Fixes #2