save all the pages
save the intermediate pages as well as the people pages.

Move the visit call into the save_page method and use that to visit all
the pages so we save them as we visit them.
Rename save_page to vist_and_save_page becuase that's what it does.
Also simplify the URL SHAing and push it out to a function to it can be
used separately from the saving.