catch errors with scraping and commit what we have so far
As the scraper is prone to falling over due to issues with the site it
is scraping wrap the mirror process in a begin/rescue and spit out a
message if it falls over. This means that it should also commit the
progress it's made to github rather than it being an all or nothing
approach.