Prototype of using ScrapedPageArchive in a decorator
This is quite rough around the edges, but it shows using
scraped_page_archive to explicitly save a response into git without
using any VCR magic.

Because we're not using VCR to serialize requests/responses anymore I've
had to add a `ResponseSerializer` class. That class outputs a slightly
different format to the current YAML because we don't have (easy) access
to the full request and we only have access to the unparsed response. I
think this is fine, since the main information we're interested in for
obtaining a response is still present.