Add MembershipList
Returns a list of Membership objects for a given Member.

The class uses capybara to visit the URL that lists the memberships of a given member. Capybara needs to visit the member's profile URL first, otherwise the membership list data will not load.

(I tried to solve this using cookies. The site sets a session id cookie when the browser visits the profile page. I sent the cookie with the page request but this did not work consistently with all membership lists.)